Created Supertags::cleanMeta and associated test.
Supertag data is always stored as an array, even if the supertag
field only accepts one value. This is so a field can change between
accepting one value and multiple values without needing to migrate
the data.

This is great, but also has one downside -- supertag fields that
only accept one value still output values in arrays. This function
takes care of popping the value out of the array.